+++ src/Node/test.js
@@ -340,20 +340,18 @@ describe('A node', () => {
 
   describe('"clone" call', () => {
 
-    it('should create a new node', () => {
-      node.merge({ stuff: true });
-      const copy = node.clone();
-
-      expect(copy).toNotBe(node);
-      expect(copy).toBeA(Node);
-      expect(toObject(copy)).toEqual(toObject(node));
-    });
-
     it('should create a node with the same ID', () => {
       const { uid } = node.meta();
-      const copy = node.clone();
+      const copy = node.new();
 
       expect(copy.meta()).toContain({ uid });
+    });
+
+    it('should not carry over any properties', () => {
+      node.merge({ original: true });
+      const copy = node.new();
+
+      expect(toObject(copy)).toEqual({});
     });
 
   });